+++ transport.c
@@ -70,11 +70,11 @@ void donor_E_dir(hydro_fields f, hydro_params p, int dir) {
 void donor_Z_dir(hydro_fields f, hydro_params p, int dir) {
 
 #ifndef SCALAR
-  double vc;
+  float vc;
   int x, y, z;
 
 
-  double ***Vbody = make_field(p);
+  float ***Vbody = make_field(p);
